Fix clippy warnings after marking some structs with `Copy`
Fix clippy warnings that raised after marking `structs` like
`Keypair`, `PrivateKey` and `Address` with the `Copy` trait.

+++ nano-zkp/examples/prover/prove.rs
@@ -36,7 +36,7 @@ fn main() {
     for i in 0..number_epochs {
         // Get random parameters.
         let (initial_pks, initial_header_hash, final_pks, block, genesis_state_commitment_opt) =
-            create_test_blocks(i as u64);
+            create_test_blocks(i);
 
         // Create genesis data.
         if i == 0 {
