+++ nano-zkp/examples/verify.rs
@@ -32,7 +32,7 @@ fn main() {
     let (initial_pks, initial_header_hash, _, _, _) = create_test_blocks(0);
 
     // Get final random parameters.
-    let (final_pks, final_header_hash, _, _, _) = create_test_blocks(number_epochs as u64);
+    let (final_pks, final_header_hash, _, _, _) = create_test_blocks(number_epochs);
 
     // Load the proof from file.
     let mut file = File::open(format!("proofs/proof_epoch_{}.bin", number_epochs)).unwrap();
